refactor dropdown component and update test
